# go-backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID           int64
	Login        string
	PasswordHash string
	CreatedAt    pgtype.Timestamptz
}

type Instrument struct {
	ID               int32
	UserID           int64
	Name             string
	ProficiencyLevel int32
}

type User struct {
	ID                        int64
	Name                      pgtype.Text
	City                      pgtype.Text
	Age                       pgtype.Int4
	Contacts                  pgtype.Text
	TheoreticalKnowledgeLevel pgtype.Int4
	HasPerformanceExperience  pgtype.Text
	AboutMe                   pgtype.Text
	ExternalLink              pgtype.Text
	IsVisible                 bool
	PhotoPath                 pgtype.Text
	AudioPath                 pgtype.Text
}

type UserGenre struct {
	ID     int32
	UserID int64
	Name   string
}

type UserLikesUser struct {
	ID           int32
	SwiperUserID int64
	TargetUserID int64
	Action       pgtype.Text
	CreatedAt    pgtype.Timestamptz
}